refactor: replace type guards with class inheritance for string instruments
