+++ src/homework_9/instruments/violin.ts
@@ -1,11 +1,8 @@
-import {BaseTunableInstrument} from '../base-tunable-instrument';
-import {IStringInstrument} from '../interfaces/i-string-instrument';
+import {BaseStringInstrument} from '../base-classes/base-string-instrument';
 
-export class Violin extends BaseTunableInstrument implements IStringInstrument {
-    private readonly strings: number = 4;
-
+export class Violin extends BaseStringInstrument {
     public constructor() {
-        super('Violin');
+        super('Violin', 4);
     }
 
     public play(): string {
@@ -13,14 +10,5 @@ export class Violin extends BaseTunableInstrument implements IStringInstrument {
             return `${this.name} is out of tune!`;
         }
         return `${this.name} creates magical melody`;
-    }
-
-    public getNumberOfStrings(): number {
-        return this.strings;
-    }
-
-    public tune(): string {
-        const baseMessage = super.tune();
-        return `${baseMessage}. All ${this.strings} strings tuned`;
     }
 }
