+++ src/homework_9/interfaces/i-tunable-instrument.ts
@@ -1,5 +1,3 @@
-import {IMusicalInstrument} from './i-musical-instrument';
-
-export interface ITunableInstrument extends IMusicalInstrument {
+export interface ITunableInstrument {
     tune(): string;
 }
